# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PR/PO Approval System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchase-requests:
    get:
      operationId: getPurchaseRequests
      tags: [purchase-requests]
      summary: Get purchase requests
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, waiting_approval, approved, rejected, completed]
        - name: type
          in: query
          schema:
            type: string
            enum: [purchase, repair, leave]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of purchase requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequestListResponse"

    post:
      operationId: createPurchaseRequest
      tags: [purchase-requests]
      summary: Create a purchase request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseRequestRequest"
      responses:
        "201":
          description: Purchase request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
  /purchase-requests/{id}:
    get:
      operationId: getPurchaseRequestById
      tags: [purchase-requests]
      summary: Get purchase request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase request found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updatePurchaseRequest
      tags: [purchase-requests]
      summary: Update a purchase request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePurchaseRequestRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
  /purchase-requests/{id}/submit:
    post:
      operationId: submitPurchaseRequest
      tags: [purchase-requests]
      summary: Submit PR for approval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
  /purchase-requests/{id}/receive:
    post:
      operationId: receivePurchaseRequest
      tags: [purchase-requests]
      summary: Mark PR as received
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReceivePRRequest"
      responses:
        "200":
          description: Received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseRequest"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    PrItem:
      type: object
      properties:
        id:
          type: integer
        prId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        qty:
          type: number
        unit:
          type: string
        estimatedPrice:
          type: number
        totalPrice:
          type: number
      required:
        - id
        - prId
        - name
        - qty
        - unit
        - estimatedPrice
        - totalPrice
    PrItemInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        qty:
          type: number
        unit:
          type: string
        estimatedPrice:
          type: number
      required:
        - name
        - qty
        - unit
        - estimatedPrice
    PurchaseRequest:
      type: object
      properties:
        id:
          type: integer
        prNumber:
          type: string
        date:
          type: string
          format: date-time
        requesterId:
          type: integer
        requesterName:
          type: string
        department:
          type: string
        type:
          type: string
          enum: [purchase, repair, leave]
        description:
          type: string
        status:
          type: string
          enum: [draft, waiting_approval, approved, rejected, completed]
        totalAmount:
          type: number
        attachmentUrl:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/PrItem"
        approvals:
          type: array
          items:
            $ref: "#/components/schemas/Approval"
        currentApprovalLevel:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - prNumber
        - date
        - requesterId
        - requesterName
        - department
        - type
        - description
        - status
        - totalAmount
        - items
        - approvals
        - createdAt
        - updatedAt
    PurchaseRequestListResponse:
      type: object
      properties:
        purchaseRequests:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseRequest"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - purchaseRequests
        - total
        - page
        - limit
    CreatePurchaseRequestRequest:
      type: object
      properties:
        type:
          type: string
          enum: [purchase, repair, leave]
        description:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PrItemInput"
        notes:
          type: string
          nullable: true
      required:
        - type
        - description
        - items
    UpdatePurchaseRequestRequest:
      type: object
      properties:
        type:
          type: string
          enum: [purchase, repair, leave]
        description:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PrItemInput"
        notes:
          type: string
          nullable: true
    ReceivePRRequest:
      type: object
      properties:
        notes:
          type: string
          nullable: true
      required: []
    Approval:
      type: object
      properties:
        id:
          type: integer
        prId:
          type: integer
        approverId:
          type: integer
        approverName:
          type: string
        level:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected]
        notes:
          type: string
          nullable: true
        actionAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - prId
        - approverId
        - approverName
        - level
        - status
        - createdAt
...